http.protocol: Support filehandlers for unhandled encodings
The setfilehandler() functions used for mime and url encoded message
bodies all operate with a signature of fh(meta, chunk, eof), but for
unhandled encodings, the callback was directly assigned to the sink
function, which has a signature of snk(chunk).  Insert a wrapper to
properly generate the EOF flag, and include a stub "meta" block
providing a virtual "name" and also the original client provided
Content-Type header, to possibly help with taking alternative
actions in the file handler.

The sink function created for raw content decoding also used the
wrong signature for the sink function.

Backport from master/for-15.05

Signed-off-by: Karl Palsson <karlp@remake.is>