dispatcher: stop at first valid child
Instead of recursing to the first valid leaf node, stop at the first
valid node at all.  Changes validity to be "having a title" and "not
being a type that recurses" rather than just allowing certain types
as well, based on discussion with jow.

This means that with some controller entries such as:
```
    entry({'home'}, call("desired_home"), _("Home"), 1)
    entry({'home', "wizard"}, firstchild(), _("Setup"), 2)
    entry({'home', 'wizard', 'initial_setup'}, call("action_wizard_home"), _("Initial Configuration"), 10)
    entry({'home', 'wizard', 'basic_view'}, view("unexepectedview"), _("Unexepected view"), 20)
```
The dispatcher will now route "cgi-bin/luci" to the
call("desired_home") rather than continuing to the "unexepectedview" as
the first leaf node of a suitable type.

Changing the list of types alone is insufficient as it would then
continue to the "call(action_wizard_home)" leaf still, instead of the
desired top level entry.

Signed-off-by: Karl Palsson <karlp@etactica.com>